- did a bit of cleanup. - disabled the runtime check in OP_CALL because as implemented it doesn't clean up properly and is not fully implemented.
